# swagger-debug.json
{"openapi":"3.0.0","paths":{"/auth/check-email":{"get":{"operationId":"AuthController_checkEmailAvailability","parameters":[{"name":"email","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["auth"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["ENTREPRENEUR","PME","ONG"]},"secteurId":{"type":"string","description":"Entrepreneur"},"paysId":{"type":"string","description":"Entrepreneur"},"domaineExpertise":{"type":"string","description":"Entrepreneur"},"objectifs":{"type":"string","description":"Entrepreneur (optionnel)"},"nomEntreprise":{"type":"string","description":"PME"},"secteurIds":{"type":"string","description":"PME — UUIDs séparés par virgules"},"nomOrganisation":{"type":"string","description":"ONG"},"domaineInterventionIds":{"type":"string","description":"ONG — UUIDs séparés par virgules"},"mission":{"type":"string","description":"ONG (optionnel)"},"logo":{"type":"string","format":"binary","description":"PME / ONG"}},"required":["email","password","role"]}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/verify-otp":{"post":{"operationId":"AuthController_verifyOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOtpDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/resend-otp":{"post":{"operationId":"AuthController_resendOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendOtpDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/reference/pays":{"get":{"operationId":"ReferenceController_getPays","parameters":[],"responses":{"200":{"description":""}},"tags":["reference"]}},"/reference/secteurs":{"get":{"operationId":"ReferenceController_getSecteurs","parameters":[],"responses":{"200":{"description":""}},"tags":["reference"]}},"/reference/domaines-intervention":{"get":{"operationId":"ReferenceController_getDomainesIntervention","parameters":[],"responses":{"200":{"description":""}},"tags":["reference"]}},"/admin/users":{"get":{"operationId":"AdminUsersController_listUsers","parameters":[{"name":"role","required":false,"in":"query","schema":{"type":"string","enum":["ENTREPRENEUR","PME","ONG","ADMINISTRATEUR"]}},{"name":"statut","required":false,"in":"query","schema":{"type":"string","enum":["EN_ATTENTE","ACTIF","SUSPENDU"]}},{"name":"search","required":false,"in":"query","description":"Recherche par email (partielle)","schema":{"type":"string"}},{"name":"inclureSupprimes","required":false,"in":"query","description":"Inclure les comptes supprimés (soft delete)","schema":{"default":false,"type":"boolean"}},{"name":"page","required":false,"in":"query","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"default":20,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin"]}},"/admin/users/{id}":{"get":{"operationId":"AdminUsersController_getUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin"]},"delete":{"operationId":"AdminUsersController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin"]}},"/admin/users/{id}/statut":{"patch":{"operationId":"AdminUsersController_updateStatut","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserStatutDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin"]}},"/admin/users/{id}/restore":{"post":{"operationId":"AdminUsersController_restoreUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["admin"]}}},"info":{"title":"Gaynaako Opportunity Agent API","description":"API de la plateforme de centralisation d'opportunités assistée par IA","version":"1.0","contact":{}},"tags":[{"name":"auth","description":"Authentification & gestion des comptes"},{"name":"users","description":"Profils utilisateurs"},{"name":"opportunities","description":"Opportunités"}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"VerifyOtpDto":{"type":"object","properties":{"email":{"type":"string"},"code":{"type":"string","example":"123456"}},"required":["email","code"]},"ResendOtpDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"ForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"ResetPasswordDto":{"type":"object","properties":{"email":{"type":"string"},"code":{"type":"string","example":"123456"},"newPassword":{"type":"string"}},"required":["email","code","newPassword"]},"UpdateUserStatutDto":{"type":"object","properties":{"statut":{"type":"string","enum":["EN_ATTENTE","ACTIF","SUSPENDU"]}},"required":["statut"]}}}}